Fix navigation and backstack problems
Check if the search view is iconified or current destination is bottom nav, then avoid navigation to search fragment.

Move the `searchFragment` navigation to `onMenuItemActionExpand` so every time search expand, try to get the search fragment.

Clear the focus of `searchView` when a query is submitted. Helps in decreasing the number of back presses to minimize player and navigate back.

No need to remove focus using `SearchResultFragment#onStop`, `BackPressedDispatche` can handle it.