fix: maximise player when playing new video
Fixes a regression from 12be4c716cd88fd00f413defd96a7d63292000ed,
where upon trying to play a new video, the player would stay minimised.
Additional, the playing video would be added to the queue, causing the
previously played video to be played again.
This is confusing to users, as it breaks the expectation of the player
opening when trying to play a video.