fix(SubscriptionsFragment): update last feed time to last video time
With the introduction of the local feed extraction, which only refreshes
the feed naturally once a day, it is possible for the user to update the
`lastFeedWatchedTime`, but not fetch the feed itself, leading to a
disconnect between both values.
When the feed is then later on refreshed, the caught-up-indicator is
only displayed for the videos since the last app open, and not all new
videos since the last refresh.