perf(LocalFeedRepository): speed up local feed extraction
Speeds up the local feed extraction, by fetching more feeds concurrently
and only applying a large timeout after fetching the full channels. This
is based on the idea that YouTube's RSS feeds (extracted by the generic
`FeedInfo`) have no/much less rate-limiting, likely due to the nature of
RSS readers.
A small timeout after each chunk is still applied, to be a good citizen
of the web and not overwhelm/spam the server.

Ref: https://github.com/TeamNewPipe/NewPipe/issues/11817