feat(VideoCardsAdapter): display icon for fully sponsored videos
Displays an icon on the thumbnail for fully sponsored videos. This
allows users to easily identify and avoid these videos, if they prefer
so. This works similarly to how the SponsorBlock extension works for the
website.

As the icon is recycled, it may keep it's visibility until the web request
succeeds, and thus being shown on non-sponsored videos.

For some reason the API uses a hash-prefix of 5 chars (instead of the
document/recommended 4 chars) and does not return videos otherwise. This is
following the behavior of the official client.