fix(Feed): only update `LastFeedWatchedTime` when update time is newer
Fixes an issue, where if the user switches between channel group,
already seen videos could be marked as new. This was caused by the
channel group having only older videos, which resetted the last watched
time to an older timepoint.
To fix this, the last watched time is only updated, when the new watched
time is newer than the previous watched time.