fix: backup import doesn't show any toast on success
The activity context passed to the methods of `SubscriptionHelper`
didn't live long enough to be able to show a toast message in case
of larger data (>10 channels for example). Thus we ensure the context
lives long enough by using the application context and not using
any fragment lifecycle dependent methods.