refactor(Preferences): use `MIGRATION` length as version
Uses the `MIGRATION` length to determine the appropriate migrations to be run.
This works, as opposed to Room migrations, there is no `AutoMigration`, i.e. no
version increase without a corresponding `PreferenceMigration`. By
automatically determining the preference version, we avoid issues, where a new
migration is created, but never actually run, as the version is not incremented.

Ref: https://github.com/libre-tube/LibreTube/commit/c746647041515ed8e52077753978fb099a53a543