feat: enable edge-to-edge in bottom sheets (#7574)
* feat: enable edge-to-edge in bottom sheets

The BottomSheetDialog internally adds an inset, leading to this inset being
displayed instead of the recyclerview for chapters and comments. Whilst it is
possible to remove, we need to manually handle the inset padding on the lists.

Ref: https://github.com/libre-tube/LibreTube/pull/7507#pullrequestreview-3058291871

* refactor: move bottomsheet inset removal to individual sheets

* fix: remove additional BottomSheet padding

Removes the additional BottomSheet bottom padding, which is already handled
through the inset applied by the sheet itself.