fix(LocalFeedRepository): update subscribed channels
Fixes an issue, where local subscriptions could display outdated information
(i.e. having the wrong avatar/name), as the where never updated after being
inserted into the database.
To avoid this, we use the channel info, which is fetched when refreshing the
local feed to also update the local subscriptions.