fix(video player): preview frame jumping around the screen when seeking
Fix seekbar preview frame jumping around the screen due to previous
coroutines are not being canceled when scrubbing the timebar.

UI changes:
The preview frame now moves as soon as the finger moves and a little
progress indicator is added below the preview frame to serve as a ui
feedback when the frame is being loaded.