feat(Preferences): remove codec preferences
Removes support for setting the preferred audio/video codec. This preferences can
be quite confusing to new, as well experienced users, since it's not immediately
clear which choice is optimal (not that there necessarily is one).
It also prohibited more modern codecs (such as AV1) being chosen automatically,
when the preference was not set to `all`, leading to a worse user experience

Overall it's better to let the player, which has more information, decide what
format to use.

Ref: https://github.com/libre-tube/LibreTube/issues/7850