feat: always remember playback speed
Removes the preference to set the playback speed, instead the last set playback
speed is remembered. This improves the UX by reducing the available options,
as well providing a consistent experience, as the manually set playback speed
was already remembered per player session, which is no longer recreated for
each video.

Ref: https://github.com/libre-tube/LibreTube/issues/7961
Ref: https://github.com/libre-tube/LibreTube/pull/4901