feat(Player): display chapters in timebar
Displays the chapters of a video as breaks in the progress bar of a video. This
allows users to quickly glance at the end of a chapter and makes navigation
through chapters easier.

The breaks are implemented by erasing the existing paint, which includes both
the SponsorBlock segments and the progress bar.

Closes: https://github.com/libre-tube/LibreTube/issues/4575